pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentskills"
version = "0.2.0"
description = "CLI for browsing and installing skills from anthropics/skills"
readme = "README.md"
license = "MIT"
authors = [
  { name = "Akshay Aggarwal" },
]
keywords = ["claude", "anthropic", "skills", "ai", "mcp", "agents"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
]
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/akshayaggarwal99/agentskills"
Repository = "https://github.com/akshayaggarwal99/agentskills"
Issues = "https://github.com/akshayaggarwal99/agentskills/issues"

[project.scripts]
skill = "agentskills.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["agentskills"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
